model/inference/requirements.txt
fastapi
uvicorn[standard]  
pydantic
jupyter==1.0.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy==1.23.2
pytest==7.1.2
python-dateutil==2.8.2
pytorch-lightning==1.2.8
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.4.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.9.3
gcsfs 
pandas
google-cloud-storage
google-cloud-bigquery
pytz 
datetime